// main-backend/frontend/src/generated/model/ingredient-category.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Enum representing the categories of ingredients
 */
export enum IngredientCategory {
    FreshProduce = 'FRESH_PRODUCE',
    RedMeatsAndPoultry = 'RED_MEATS_AND_POULTRY',
    Fish = 'FISH',
    DairyAndEggs = 'DAIRY_AND_EGGS',
    Bakery = 'BAKERY',
    PantryStaples = 'PANTRY_STAPLES',
    FrozenFoods = 'FROZEN_FOODS',
    Beverages = 'BEVERAGES',
    Snacks = 'SNACKS',
    Other = 'OTHER'
}