// main-backend/frontend/src/generated/model/ingredient-list-paginated-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Pageable } from './pageable';
import { IngredientDto } from './ingredient-dto';
import { Sort } from './sort';


/**
 * Paginated list containing ingredient dtos.
 */
export interface IngredientListPaginatedDto { 
    totalElements: number;
    totalPages: number;
    sort: Sort;
    first: boolean;
    last: boolean;
    number: number;
    pageable: Pageable;
    numberOfElements: number;
    size: number;
    empty: boolean;
    content?: Array<IngredientDto>;
}