// main-backend/frontend/src/generated/model/shopping-list-ingredient-add-dto.ts
/**
 * merged spec
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { IngredientUnitDto } from './ingredient-unit-dto';


export interface ShoppingListIngredientAddDto { 
    /**
     * Unique id of the ingredient to be added. Used only when we add an ingredient existing in the database to the shopping list.
     */
    existingIngredientId?: number;
    /**
     * Name of the ingredient to be added. Used only when we add a ingredient not existing in the database to the shopping list. 
     */
    customIngredientName?: string;
    unit: IngredientUnitDto;
    /**
     * Amount of the ingredient used in regard to the unit.
     */
    amount: number;
}